+++ src/main/java/ford/group/orderapp/dto/ordereditem/OrderedItemToSaveDTO.java
@@ -4,7 +4,7 @@ import ford.group.orderapp.dto.order.OrderDTO;
 import ford.group.orderapp.dto.product.ProductDTO;
 
 public record OrderedItemToSaveDTO(OrderDTO order,
-                                    ProductDTO product,
-                                    Long requestedAmount,
-                                    Float unitPrice) {
+                                   ProductDTO product,
+                                   Long requestedAmount,
+                                   Double unitPrice) {
 }
